src/third_party/blink/web_tests/external/wpt/css/css-ui/parsing/user-select-computed-expected.txt

This is a testharness.js-based test.
PASS Property user-select value 'auto' computes to 'auto'
PASS Property user-select value 'text' computes to 'text'
PASS Property user-select value 'none' computes to 'none'
FAIL Property user-select value 'contain' computes to 'contain' assert_equals: expected "contain" but got "auto"
PASS Property user-select value 'all' computes to 'all'
Harness: the test ran to completion.
